+++ src/main/java/com/spring/java/backend/repositories/FormaPagamentoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.spring.java.backend.entities.FormaPagamento;
 
 @Repository
-public interface FormaPagamentoRepository extends JpaRepository<FormaPagamento, Long>{
+public interface FormaPagamentoRepository extends JpaRepository<FormaPagamento, Integer>{
 
 }
